AP_RangeFinder: adjust signature of report_malfunction
old packet is 8 bits, new packet is 16; we don't use any but the first 8 so trunction is reasonable; this change makes it promote the uint8_t.

+++ libraries/AP_RangeFinder/AP_RangeFinder_Ainstein_LR_D1.cpp
@@ -173,7 +173,7 @@ bool AP_RangeFinder_Ainstein_LR_D1::get_one_reading(float &reading_m)
 }
 
 #if AP_RANGEFINDER_AINSTEIN_LR_D1_SHOW_MALFUNCTIONS
-void AP_RangeFinder_Ainstein_LR_D1::report_malfunction(const uint8_t _malfunction_alert_, const uint8_t _malfunction_alert_prev_)
+void AP_RangeFinder_Ainstein_LR_D1::report_malfunction(const uint16_t _malfunction_alert_, const uint16_t _malfunction_alert_prev_)
 {
     static const struct {
         MalfunctionAlert bit;
